Move AuthnRequest to separate file
